+++ spec/listen/listener_spec.rb
@@ -61,13 +61,46 @@ describe Listen::Listener do
     end
   end
 
-  describe '#stop' do
-    it "stops adapter" do
+  context "with a started listener" do
+    before do
       subject.stub(:initialize_adapter) { adapter }
       subject.stub(:init_paths)
       subject.start
-      adapter.should_receive(:stop)
-      subject.stop
+    end
+
+    describe '#stop' do
+      it "stops adapter" do
+        adapter.should_receive(:stop)
+        subject.stop
+      end
+
+      it "unpause if paused" do
+        subject.paused = true
+        adapter.should_not_receive(:stop)
+        subject.stop
+        subject.paused.should be_false
+      end
+    end
+
+    describe "#pause" do
+      it "stops adapter and wait until unpaused" do
+        adapter.should_receive(:stop)
+        subject.should_receive(:sleep).any_number_of_times
+        Thread.new { subject.pause }
+        sleep 0.0001
+        subject.paused.should be_true
+        subject.paused = false
+      end
+    end
+
+    describe "#unpause" do
+      it "unpauses and starts adapter" do
+        Thread.new { subject.pause }
+        subject.should_receive(:init_paths)
+        adapter.should_receive(:start)
+        subject.unpause
+        subject.paused.should be_false
+      end
     end
   end